+++ matriz.hpp
@@ -3,14 +3,14 @@ class Matriz {
 		//Establecer = Set
 	public:
 		//-------- setters y getters --------
-		int getFila();
-		void setFila ( int filaPublica );
-		int getColumna();
-		void setColumna ( int columnaPublica );
-		int getMatriz();
-		void setMatriz ( int **matrizPublica );
-		int getMatrizResultado();
-		void setMatrizResultado ( int **matrizPublica );
+		unsigned int getFila();
+		void setFila ( unsigned int filaPublica );
+		unsigned int getColumna();
+		void setColumna ( unsigned int columnaPublica );
+		double getMatriz();
+		void setMatriz ( double **matrizPublica );
+		double getMatrizResultado();
+		void setMatrizResultado ( double **matrizPublica );
 		//-------- trabajo con archivos --------
 		void leerArchivo();
 		void guardarEnArchivo();
@@ -20,16 +20,16 @@ class Matriz {
 		void editarMatrizVacia();
 		void mostrarMatriz();
 		void mostrarMatrizResultado();
-		int traza();
-		int determinante();
+		double traza();
+		double determinante();
 		void traspuesta();
 		void adjunta();
 		void triangular();
 		void inversa();
 		void gauss();
 	private:
-		int fila;
-		int columna;
-		int **matriz;
-		int **matrizResultado;
+		unsigned int fila;
+		unsigned int columna;
+		double **matriz;
+		double **matrizResultado;
 	};
